requirements.txt
tensorflow-gpu==1.14
numpy==1.14.5
Pillow
requests
keras==2.3.1
dlib
tqdm
Flask==0.12.2
gunicorn==19.9.0